Setting up database schema for mosques, families and sectretaries (#4)
In this commit I have added the schema for the above mentioned tables. The relationship is meant to be as follows.
A mosque can have many families. A secreteary of the mosque has a direct relationship to the family as well as to the mosque.
Finally added some unit tests for the model. WIll have to add more unit tests.